pyproject.toml
[project]
name = "adk-mcp-ping"
version = "0.1.0"
description = "Keep-alive pings for Google ADK MCP connections"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Stephen Matta"}
]
keywords = ["adk", "mcp", "google", "agent", "keep-alive", "ping"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "google-adk>=0.1.0",
    "mcp>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "deptry>=0.21.0",
    "pre-commit>=4.0.0",
    "poethepoet>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adk_mcp_ping"]

# Poethepoet task runner configuration
[tool.poe.tasks]
# Linting and formatting
lint = "ruff check src tests"
lint-fix = "ruff check --fix src tests"
format = "ruff format src tests"
format-check = "ruff format --check src tests"

# Type checking
typecheck = "mypy src/adk_mcp_ping"

# Dependency checking
deps = "deptry src"

# Testing
test = "pytest"
test-cov = "pytest --cov --cov-report=term-missing --cov-report=html"
test-fast = "pytest -x --tb=short"

# Pre-commit
hooks-run = "pre-commit run --all-files"

# Combined checks
check = ["lint", "format-check", "typecheck"]
fix = ["lint-fix", "format"]

# Table-style tasks (must come after inline tasks)
[tool.poe.tasks.hooks-install]
shell = "pre-commit install && pre-commit install --hook-type pre-push"
help = "Install pre-commit and pre-push hooks"

[tool.poe.tasks.ci]
sequence = ["check", "deps", "test"]
help = "Run all CI checks"

[tool.poe.tasks.dev]
sequence = ["fix", "test-fast"]
help = "Fix issues and run quick tests"

[tool.poe.tasks.qa]
sequence = ["fix", "typecheck", "deps", "test-cov"]
help = "Full quality assurance check"

# Pytest configuration
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/adk_mcp_ping"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# Ruff configuration
[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented out code)
    "PL",     # pylint
    "RUF",    # ruff-specific
]
ignore = [
    "PLR0913",  # too many arguments
    "PLR2004",  # magic value comparison
]

[tool.ruff.lint.isort]
known-first-party = ["adk_mcp_ping"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004", "PLC0415", "E501", "SIM105", "TC003"]

# Mypy configuration
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# Deptry configuration
[tool.deptry]
extend_exclude = ["tests"]

[tool.deptry.per_rule_ignores]
DEP002 = ["google-adk", "mcp"]  # These are used but deptry may not detect dynamic imports
